Create admin package for tests + beginning of tests
