+++ app/src/androidTest/java/com/github/sdpteam15/polyevents/admin/AdminHubFragmentTest.kt
@@ -1,4 +1,4 @@
-package com.github.sdpteam15.polyevents
+package com.github.sdpteam15.polyevents.admin
 
 import android.content.Intent
 import androidx.test.core.app.ActivityScenario
@@ -11,14 +11,15 @@ import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.sdpteam15.polyevents.admin.ZoneManagementActivity
+import com.github.sdpteam15.polyevents.MainActivity
+import com.github.sdpteam15.polyevents.R
 import com.github.sdpteam15.polyevents.database.Database
 import com.github.sdpteam15.polyevents.database.DatabaseInterface
 import com.github.sdpteam15.polyevents.database.FirestoreDatabaseProvider
+import com.github.sdpteam15.polyevents.fakedatabase.FakeDatabase
 import com.github.sdpteam15.polyevents.login.UserLogin
 import com.github.sdpteam15.polyevents.model.UserEntity
 import com.github.sdpteam15.polyevents.model.UserProfile
-import com.google.firebase.auth.FirebaseAuth
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
@@ -27,7 +28,7 @@ import org.mockito.Mockito.mock
 import org.mockito.Mockito.`when` as When
 
 @RunWith(AndroidJUnit4::class)
-class ZoneManagementListTest {
+class AdminHubFragmentTest {
     var mainActivity = ActivityScenarioRule(MainActivity::class.java)
     lateinit var scenario: ActivityScenario<MainActivity>
 
@@ -37,14 +38,12 @@ class ZoneManagementListTest {
     val uid = "testUid"
     val username = "JohnDoe"
     val email = "John@Doe.com"
-    lateinit var mockedDatabase: DatabaseInterface
 
     @Before
     fun setup() {
-        mockedDatabase = mock(DatabaseInterface::class.java)
+        val mockedDatabase = mock(DatabaseInterface::class.java)
         val mockedUserProfile = UserProfile("TestID", "TestName")
         When(mockedDatabase.currentProfile).thenReturn(mockedUserProfile)
-
         Database.currentDatabase = mockedDatabase
 
         UserLogin.currentUserLogin.signOut()
@@ -57,10 +56,10 @@ class ZoneManagementListTest {
 
         val intent = Intent(ApplicationProvider.getApplicationContext(), MainActivity::class.java)
         scenario = ActivityScenario.launch(intent)
+
         Espresso.onView(ViewMatchers.withId(R.id.ic_home)).perform(click())
         Espresso.onView(ViewMatchers.withId(R.id.id_fragment_admin_hub))
             .check(ViewAssertions.matches(ViewMatchers.isDisplayed()))
-        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectZoneManagement)).perform(click())
         Intents.init()
     }
 
@@ -73,15 +72,28 @@ class ZoneManagementListTest {
     }
 
     @Test
-    fun clickOnBtnCreateZoneLaunchCorrectActivityWithEmptyFields() {
-        Espresso.onView(ViewMatchers.withId(R.id.btnNewZone)).perform(click())
-        Intents.intended(IntentMatchers.hasComponent(ZoneManagementActivity::class.java.name))
+    fun clickOnBtnItemRequestManagementDisplayCorrectActivity() {
+        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectItemReqManagement)).perform(click())
+        Intents.intended(IntentMatchers.hasComponent(ItemRequestManagementActivity::class.java.name))
+    }
 
-        Espresso.onView(ViewMatchers.withId(R.id.btnManage))
-            .check(ViewAssertions.matches(ViewMatchers.withText("Create zone")))
-        Espresso.onView(ViewMatchers.withId(R.id.zoneManagementDescription))
-            .check(ViewAssertions.matches(ViewMatchers.withText("")))
-        Espresso.onView(ViewMatchers.withId(R.id.zoneManagementName))
-            .check(ViewAssertions.matches(ViewMatchers.withText("")))
+    @Test
+    fun clickOnBtnZoneManagementDisplayCorrectActivity() {
+        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectZoneManagement)).perform(click())
+        Intents.intended(IntentMatchers.hasComponent(ZoneManagementListActivity::class.java.name))
+    }
+
+    @Test
+    fun clickOnBtnEventDisplayCorrectActivity() {
+        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectEventManager)).perform(click())
+        Intents.intended(IntentMatchers.hasComponent(EventManagementActivity::class.java.name))
+    }
+
+    @Test
+    fun clickOnBtnUserManagementDisplayCorrectActivity() {
+        Database.currentDatabase = FakeDatabase
+        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectUserManagement)).perform(click())
+        Intents.intended(IntentMatchers.hasComponent(UserManagementListActivity::class.java.name))
+        Database.currentDatabase = FirestoreDatabaseProvider
     }
 }
