+++ app/src/androidTest/java/com/github/sdpteam15/polyevents/admin/UserManagementListTest.kt
@@ -1,0 +1,5 @@
+package com.github.sdpteam15.polyevents.admin
+
+class UserManagementListTest {
+
+}
