+++ app/src/androidTest/java/com/github/sdpteam15/polyevents/admin/ZoneManagementListTest.kt
@@ -1,4 +1,4 @@
-package com.github.sdpteam15.polyevents
+package com.github.sdpteam15.polyevents.admin
 
 import android.content.Intent
 import androidx.test.core.app.ActivityScenario
@@ -11,26 +11,23 @@ import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.sdpteam15.polyevents.admin.*
+import com.github.sdpteam15.polyevents.MainActivity
+import com.github.sdpteam15.polyevents.R
 import com.github.sdpteam15.polyevents.database.Database
 import com.github.sdpteam15.polyevents.database.DatabaseInterface
 import com.github.sdpteam15.polyevents.database.FirestoreDatabaseProvider
-import com.github.sdpteam15.polyevents.fakedatabase.FakeDatabase
 import com.github.sdpteam15.polyevents.login.UserLogin
 import com.github.sdpteam15.polyevents.model.UserEntity
 import com.github.sdpteam15.polyevents.model.UserProfile
-import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.ktx.Firebase
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.mock
-import org.mockito.kotlin.anyOrNull
 import org.mockito.Mockito.`when` as When
 
 @RunWith(AndroidJUnit4::class)
-class AdminHubFragmentTest {
+class ZoneManagementListTest {
     var mainActivity = ActivityScenarioRule(MainActivity::class.java)
     lateinit var scenario: ActivityScenario<MainActivity>
 
@@ -40,12 +37,14 @@ class AdminHubFragmentTest {
     val uid = "testUid"
     val username = "JohnDoe"
     val email = "John@Doe.com"
+    lateinit var mockedDatabase: DatabaseInterface
 
     @Before
     fun setup() {
-        val mockedDatabase = mock(DatabaseInterface::class.java)
+        mockedDatabase = mock(DatabaseInterface::class.java)
         val mockedUserProfile = UserProfile("TestID", "TestName")
         When(mockedDatabase.currentProfile).thenReturn(mockedUserProfile)
+
         Database.currentDatabase = mockedDatabase
 
         UserLogin.currentUserLogin.signOut()
@@ -58,10 +57,10 @@ class AdminHubFragmentTest {
 
         val intent = Intent(ApplicationProvider.getApplicationContext(), MainActivity::class.java)
         scenario = ActivityScenario.launch(intent)
-
         Espresso.onView(ViewMatchers.withId(R.id.ic_home)).perform(click())
         Espresso.onView(ViewMatchers.withId(R.id.id_fragment_admin_hub))
             .check(ViewAssertions.matches(ViewMatchers.isDisplayed()))
+        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectZoneManagement)).perform(click())
         Intents.init()
     }
 
@@ -74,28 +73,15 @@ class AdminHubFragmentTest {
     }
 
     @Test
-    fun clickOnBtnItemRequestManagementDisplayCorrectActivity() {
-        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectItemReqManagement)).perform(click())
-        Intents.intended(IntentMatchers.hasComponent(ItemRequestManagementActivity::class.java.name))
-    }
+    fun clickOnBtnCreateZoneLaunchCorrectActivityWithEmptyFields() {
+        Espresso.onView(ViewMatchers.withId(R.id.btnNewZone)).perform(click())
+        Intents.intended(IntentMatchers.hasComponent(ZoneManagementActivity::class.java.name))
 
-    @Test
-    fun clickOnBtnZoneManagementDisplayCorrectActivity() {
-        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectZoneManagement)).perform(click())
-        Intents.intended(IntentMatchers.hasComponent(ZoneManagementListActivity::class.java.name))
-    }
-
-    @Test
-    fun clickOnBtnEventDisplayCorrectActivity() {
-        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectEventManager)).perform(click())
-        Intents.intended(IntentMatchers.hasComponent(EventManagementActivity::class.java.name))
-    }
-
-    @Test
-    fun clickOnBtnUserManagementDisplayCorrectActivity() {
-        Database.currentDatabase = FakeDatabase
-        Espresso.onView(ViewMatchers.withId(R.id.btnRedirectUserManagement)).perform(click())
-        Intents.intended(IntentMatchers.hasComponent(UserManagementListActivity::class.java.name))
-        Database.currentDatabase = FirestoreDatabaseProvider
+        Espresso.onView(ViewMatchers.withId(R.id.btnManage))
+            .check(ViewAssertions.matches(ViewMatchers.withText("Create zone")))
+        Espresso.onView(ViewMatchers.withId(R.id.zoneManagementDescription))
+            .check(ViewAssertions.matches(ViewMatchers.withText("")))
+        Espresso.onView(ViewMatchers.withId(R.id.zoneManagementName))
+            .check(ViewAssertions.matches(ViewMatchers.withText("")))
     }
 }
